requirements.txt
fastapi
uvicorn[standard]
aiosqlite
httpx
mcp[cli]
pyyaml
pydantic
pydantic-settings
